Move utils to save_features_labels and fix type hints
